green: Add InputContainer
