+++ src/App.jsx
@@ -1,30 +1,9 @@
-import { useDispatch, useSelector } from 'react-redux';
-
-import { updateTaskTitle, addTask } from './actions';
+import React from 'react';
 
 import Page from './Page';
 
 export default function App() {
-  const { taskTitle } = useSelector((state) => ({
-    taskTitle: state.taskTitle,
-  }));
-
-  const dispatch = useDispatch();
-
-  function handleChangeTitle(event) {
-    dispatch(updateTaskTitle(event.target.value));
-  }
-
-  function handleClickAddTask() {
-    dispatch(addTask());
-  }
-
   return (
-    <Page
-      taskTitle={taskTitle}
-      onChangeTitle={handleChangeTitle}
-      onClickAddTask={handleClickAddTask}
-
-    />
+    <Page />
   );
 }
